producer/requirements.txt
flask
kafka-python